requirements.txt
# Apache Beam
apache-beam[gcp]==2.52.0

# Kafka
confluent-kafka==2.3.0

# Data processing
polars==0.20.3

# MongoDB
pymongo==4.6.1

# Configuration
pyyaml==6.0.1

# Logging and utilities
python-dateutil==2.8.2
